+++ test/mocks/plugin-selecter2.js
@@ -3,7 +3,7 @@ const Plugin = require('../../src/plugins/Plugin.js')
 export default class TestSelector extends Plugin {
   constructor (core, opts) {
     super(core, opts)
-    this.type = 'selecter'
+    this.type = 'selecter2'
   }
 
   run (results) {